+++ Editor/NDMF/Passes.CloningMaterialsPass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: de8903e36e3955f4d80ae9032de8d7da
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
