+++ Editor/NDMF/Passes.FinalizePass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ac3ff6e71ed26d848a7ec480ce50d77a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
