+++ Editor/NDMF/Passes.GenerateAdditionalControlPass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 05385a5c8d3d01943a25236828135544
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
