+++ Editor/NDMF/Passes.GenerateAnimationsPass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ff059391638cff049aafe19301f32577
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
