+++ Editor/NDMF/Passes.NormalizeMaterialsPass.cs
@@ -1,0 +1,23 @@
+﻿using System.Linq;
+using nadena.dev.ndmf;
+using UnityEngine;
+
+namespace io.github.azukimochi
+{
+    partial class Passes
+    {
+        internal sealed class NormalizeMaterialsPass : LightLimitChangerBasePass<NormalizeMaterialsPass>
+        {
+            protected override void Execute(BuildContext context, Session session, LightLimitChangerObjectCache cache)
+            {
+                foreach (var material in cache.MappedObjects.Select(x => x as Material).Where(x => x != null).ToArray())
+                {
+                    if (ShaderInfo.TryGetShaderInfo(material, out var shaderInfo))
+                    {
+                        shaderInfo.TryNormalizeMaterial(material, cache);
+                    }
+                }
+            }
+        }
+    }
+}
