+++ Editor/NDMF/Passes.NormalizeMaterialsPass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e5871e424abf8264e972b8663bd38fd1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
